Restructure widgets, add minimum padding
