+++ website/pubspec.yaml
@@ -7,6 +7,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  url_launcher:
   super_editor:
     git:
       url: https://github.com/superlistapp/super_editor.git
